@@ -13,9 +13,9 @@ include(ExternalProject)
 
 set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 
-CHECK_CXX_COMPILER_FLAG("-std=c++11" CXX11)
-if (NOT CXX11)
-    message(FATAL_ERROR "C++ compiler doesn't support C++11")
+CHECK_CXX_COMPILER_FLAG("-std=c++14" CXX14)
+if (NOT CXX14)
+    message(FATAL_ERROR "C++ compiler doesn't support C++14")
 endif()
 
 CHECK_INCLUDE_FILES("inttypes.h" HAVE_INTTYPES_H)
@@ -307,4 +307,4 @@ add_executable(
 )
 add_dependencies(benchmark thrift msgpack protobuf capnproto boost cereal avro flatbuffers yas)
 target_link_libraries(benchmark ${LINKLIBS})
-set_target_properties(benchmark PROPERTIES COMPILE_FLAGS "-std=c++11")
+set_target_properties(benchmark PROPERTIES COMPILE_FLAGS "-std=c++14")
