@@ -6,7 +6,6 @@
 #include <chrono>
 #include <sstream>
 
-#include <boost/shared_ptr.hpp>
 #include <boost/lexical_cast.hpp>
 
 #include <thrift/transport/TBufferTransports.h>
@@ -43,8 +42,8 @@ thrift_serialization_test(size_t iterations, ThriftSerializationProto proto = Th
 
     using namespace thrift_test;
 
-    boost::shared_ptr<TMemoryBuffer> buffer1(new TMemoryBuffer());
-    boost::shared_ptr<TMemoryBuffer> buffer2(new TMemoryBuffer());
+    std::shared_ptr<TMemoryBuffer> buffer1(new TMemoryBuffer());
+    std::shared_ptr<TMemoryBuffer> buffer2(new TMemoryBuffer());
 
     TBinaryProtocolT<TMemoryBuffer> binary_protocol1(buffer1);
     TBinaryProtocolT<TMemoryBuffer> binary_protocol2(buffer2);
@@ -95,7 +94,7 @@ thrift_serialization_test(size_t iterations, ThriftSerializationProto proto = Th
         tag = "thrift-compact:";
     }
 
-    std::cout << tag << " version = " << VERSION << std::endl;
+    std::cout << tag << " version = " << PACKAGE_VERSION << std::endl;
     std::cout << tag << " size = " << serialized.size() << " bytes" << std::endl;
 
     auto start = std::chrono::high_resolution_clock::now();
