@@ -1,127 +1,100 @@
 // Generated by the protocol buffer compiler.  DO NOT EDIT!
 // source: test.proto
 
-#define INTERNAL_SUPPRESS_PROTOBUF_FIELD_DEPRECATION
 #include "test.pb.h"
 
 #include <algorithm>
 
 #include <google/protobuf/stubs/common.h>
-#include <google/protobuf/stubs/port.h>
-#include <google/protobuf/stubs/once.h>
 #include <google/protobuf/io/coded_stream.h>
+#include <google/protobuf/extension_set.h>
 #include <google/protobuf/wire_format_lite_inl.h>
 #include <google/protobuf/descriptor.h>
 #include <google/protobuf/generated_message_reflection.h>
 #include <google/protobuf/reflection_ops.h>
 #include <google/protobuf/wire_format.h>
 // @@protoc_insertion_point(includes)
+#include <google/protobuf/port_def.inc>
 
 namespace protobuf_test {
-
-namespace {
-
-const ::google::protobuf::Descriptor* Record_descriptor_ = NULL;
-const ::google::protobuf::internal::GeneratedMessageReflection*
-  Record_reflection_ = NULL;
-
-}  // namespace
-
-
-void protobuf_AssignDesc_test_2eproto() GOOGLE_ATTRIBUTE_COLD;
-void protobuf_AssignDesc_test_2eproto() {
-  protobuf_AddDesc_test_2eproto();
-  const ::google::protobuf::FileDescriptor* file =
-    ::google::protobuf::DescriptorPool::generated_pool()->FindFileByName(
-      "test.proto");
-  GOOGLE_CHECK(file != NULL);
-  Record_descriptor_ = file->message_type(0);
-  static const int Record_offsets_[2] = {
-    GOOGLE_PROTOBUF_GENERATED_MESSAGE_FIELD_OFFSET(Record, ids_),
-    GOOGLE_PROTOBUF_GENERATED_MESSAGE_FIELD_OFFSET(Record, strings_),
-  };
-  Record_reflection_ =
-    ::google::protobuf::internal::GeneratedMessageReflection::NewGeneratedMessageReflection(
-      Record_descriptor_,
-      Record::internal_default_instance(),
-      Record_offsets_,
-      -1,
-      -1,
-      -1,
-      sizeof(Record),
-      GOOGLE_PROTOBUF_GENERATED_MESSAGE_FIELD_OFFSET(Record, _internal_metadata_));
-}
-
-namespace {
-
-GOOGLE_PROTOBUF_DECLARE_ONCE(protobuf_AssignDescriptors_once_);
-void protobuf_AssignDescriptorsOnce() {
-  ::google::protobuf::GoogleOnceInit(&protobuf_AssignDescriptors_once_,
-                 &protobuf_AssignDesc_test_2eproto);
-}
-
-void protobuf_RegisterTypes(const ::std::string&) GOOGLE_ATTRIBUTE_COLD;
-void protobuf_RegisterTypes(const ::std::string&) {
-  protobuf_AssignDescriptorsOnce();
-  ::google::protobuf::MessageFactory::InternalRegisterGeneratedMessage(
-      Record_descriptor_, Record::internal_default_instance());
-}
-
-}  // namespace
-
-void protobuf_ShutdownFile_test_2eproto() {
-  Record_default_instance_.Shutdown();
-  delete Record_reflection_;
-}
-
-void protobuf_InitDefaults_test_2eproto_impl() {
+class RecordDefaultTypeInternal {
+ public:
+  ::google::protobuf::internal::ExplicitlyConstructed<Record> _instance;
+} _Record_default_instance_;
+}  // namespace protobuf_test
+static void InitDefaultsRecord_test_2eproto() {
   GOOGLE_PROTOBUF_VERIFY_VERSION;
 
-  ::google::protobuf::internal::GetEmptyString();
-  Record_default_instance_.DefaultConstruct();
-  Record_default_instance_.get_mutable()->InitAsDefaultInstance();
-}
-
-GOOGLE_PROTOBUF_DECLARE_ONCE(protobuf_InitDefaults_test_2eproto_once_);
-void protobuf_InitDefaults_test_2eproto() {
-  ::google::protobuf::GoogleOnceInit(&protobuf_InitDefaults_test_2eproto_once_,
-                 &protobuf_InitDefaults_test_2eproto_impl);
-}
-void protobuf_AddDesc_test_2eproto_impl() {
-  GOOGLE_PROTOBUF_VERIFY_VERSION;
-
-  protobuf_InitDefaults_test_2eproto();
-  ::google::protobuf::DescriptorPool::InternalAddGeneratedFile(
-    "\n\ntest.proto\022\rprotobuf_test\"&\n\006Record\022\013\n"
-    "\003ids\030\001 \003(\003\022\017\n\007strings\030\002 \003(\tb\006proto3", 75);
-  ::google::protobuf::MessageFactory::InternalRegisterGeneratedFile(
-    "test.proto", &protobuf_RegisterTypes);
-  ::google::protobuf::internal::OnShutdown(&protobuf_ShutdownFile_test_2eproto);
-}
-
-GOOGLE_PROTOBUF_DECLARE_ONCE(protobuf_AddDesc_test_2eproto_once_);
-void protobuf_AddDesc_test_2eproto() {
-  ::google::protobuf::GoogleOnceInit(&protobuf_AddDesc_test_2eproto_once_,
-                 &protobuf_AddDesc_test_2eproto_impl);
-}
-// Force AddDescriptors() to be called at static initialization time.
-struct StaticDescriptorInitializer_test_2eproto {
-  StaticDescriptorInitializer_test_2eproto() {
-    protobuf_AddDesc_test_2eproto();
+  {
+    void* ptr = &::protobuf_test::_Record_default_instance_;
+    new (ptr) ::protobuf_test::Record();
+    ::google::protobuf::internal::OnShutdownDestroyMessage(ptr);
   }
-} static_descriptor_initializer_test_2eproto_;
-
-namespace {
-
-static void MergeFromFail(int line) GOOGLE_ATTRIBUTE_COLD GOOGLE_ATTRIBUTE_NORETURN;
-static void MergeFromFail(int line) {
-  ::google::protobuf::internal::MergeFromFail(__FILE__, line);
+  ::protobuf_test::Record::InitAsDefaultInstance();
 }
 
-}  // namespace
+::google::protobuf::internal::SCCInfo<0> scc_info_Record_test_2eproto =
+    {{ATOMIC_VAR_INIT(::google::protobuf::internal::SCCInfoBase::kUninitialized), 0, InitDefaultsRecord_test_2eproto}, {}};
 
+void InitDefaults_test_2eproto() {
+  ::google::protobuf::internal::InitSCC(&scc_info_Record_test_2eproto.base);
+}
+
+::google::protobuf::Metadata file_level_metadata_test_2eproto[1];
+constexpr ::google::protobuf::EnumDescriptor const** file_level_enum_descriptors_test_2eproto = nullptr;
+constexpr ::google::protobuf::ServiceDescriptor const** file_level_service_descriptors_test_2eproto = nullptr;
+
+const ::google::protobuf::uint32 TableStruct_test_2eproto::offsets[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) = {
+  ~0u,  // no _has_bits_
+  PROTOBUF_FIELD_OFFSET(::protobuf_test::Record, _internal_metadata_),
+  ~0u,  // no _extensions_
+  ~0u,  // no _oneof_case_
+  ~0u,  // no _weak_field_map_
+  PROTOBUF_FIELD_OFFSET(::protobuf_test::Record, ids_),
+  PROTOBUF_FIELD_OFFSET(::protobuf_test::Record, strings_),
+};
+static const ::google::protobuf::internal::MigrationSchema schemas[] PROTOBUF_SECTION_VARIABLE(protodesc_cold) = {
+  { 0, -1, sizeof(::protobuf_test::Record)},
+};
+
+static ::google::protobuf::Message const * const file_default_instances[] = {
+  reinterpret_cast<const ::google::protobuf::Message*>(&::protobuf_test::_Record_default_instance_),
+};
+
+::google::protobuf::internal::AssignDescriptorsTable assign_descriptors_table_test_2eproto = {
+  {}, AddDescriptors_test_2eproto, "test.proto", schemas,
+  file_default_instances, TableStruct_test_2eproto::offsets,
+  file_level_metadata_test_2eproto, 1, file_level_enum_descriptors_test_2eproto, file_level_service_descriptors_test_2eproto,
+};
+
+const char descriptor_table_protodef_test_2eproto[] =
+  "\n\ntest.proto\022\rprotobuf_test\"&\n\006Record\022\013\n"
+  "\003ids\030\001 \003(\003\022\017\n\007strings\030\002 \003(\tb\006proto3"
+  ;
+::google::protobuf::internal::DescriptorTable descriptor_table_test_2eproto = {
+  false, InitDefaults_test_2eproto, 
+  descriptor_table_protodef_test_2eproto,
+  "test.proto", &assign_descriptors_table_test_2eproto, 75,
+};
+
+void AddDescriptors_test_2eproto() {
+  static constexpr ::google::protobuf::internal::InitFunc deps[1] =
+  {
+  };
+ ::google::protobuf::internal::AddDescriptors(&descriptor_table_test_2eproto, deps, 0);
+}
+
+// Force running AddDescriptors() at dynamic initialization time.
+static bool dynamic_init_dummy_test_2eproto = []() { AddDescriptors_test_2eproto(); return true; }();
+namespace protobuf_test {
 
 // ===================================================================
+
+void Record::InitAsDefaultInstance() {
+}
+class Record::HasBitSetters {
+ public:
+};
 
 #if !defined(_MSC_VER) || _MSC_VER >= 1900
 const int Record::kIdsFieldNumber;
@@ -129,25 +102,22 @@ const int Record::kStringsFieldNumber;
 #endif  // !defined(_MSC_VER) || _MSC_VER >= 1900
 
 Record::Record()
-  : ::google::protobuf::Message(), _internal_metadata_(NULL) {
-  if (this != internal_default_instance()) protobuf_InitDefaults_test_2eproto();
+  : ::google::protobuf::Message(), _internal_metadata_(nullptr) {
   SharedCtor();
   // @@protoc_insertion_point(constructor:protobuf_test.Record)
 }
-
-void Record::InitAsDefaultInstance() {
-}
-
 Record::Record(const Record& from)
   : ::google::protobuf::Message(),
-    _internal_metadata_(NULL) {
-  SharedCtor();
-  UnsafeMergeFrom(from);
+      _internal_metadata_(nullptr),
+      ids_(from.ids_),
+      strings_(from.strings_) {
+  _internal_metadata_.MergeFrom(from._internal_metadata_);
   // @@protoc_insertion_point(copy_constructor:protobuf_test.Record)
 }
 
 void Record::SharedCtor() {
-  _cached_size_ = 0;
+  ::google::protobuf::internal::InitSCC(
+      &scc_info_Record_test_2eproto.base);
 }
 
 Record::~Record() {
@@ -159,90 +129,150 @@ void Record::SharedDtor() {
 }
 
 void Record::SetCachedSize(int size) const {
-  GOOGLE_SAFE_CONCURRENT_WRITES_BEGIN();
-  _cached_size_ = size;
-  GOOGLE_SAFE_CONCURRENT_WRITES_END();
+  _cached_size_.Set(size);
 }
-const ::google::protobuf::Descriptor* Record::descriptor() {
-  protobuf_AssignDescriptorsOnce();
-  return Record_descriptor_;
-}
-
 const Record& Record::default_instance() {
-  protobuf_InitDefaults_test_2eproto();
+  ::google::protobuf::internal::InitSCC(&::scc_info_Record_test_2eproto.base);
   return *internal_default_instance();
 }
 
-::google::protobuf::internal::ExplicitlyConstructed<Record> Record_default_instance_;
-
-Record* Record::New(::google::protobuf::Arena* arena) const {
-  Record* n = new Record;
-  if (arena != NULL) {
-    arena->Own(n);
-  }
-  return n;
-}
 
 void Record::Clear() {
 // @@protoc_insertion_point(message_clear_start:protobuf_test.Record)
+  ::google::protobuf::uint32 cached_has_bits = 0;
+  // Prevent compiler warnings about cached_has_bits being unused
+  (void) cached_has_bits;
+
   ids_.Clear();
   strings_.Clear();
+  _internal_metadata_.Clear();
 }
 
+#if GOOGLE_PROTOBUF_ENABLE_EXPERIMENTAL_PARSER
+const char* Record::_InternalParse(const char* begin, const char* end, void* object,
+                  ::google::protobuf::internal::ParseContext* ctx) {
+  auto msg = static_cast<Record*>(object);
+  ::google::protobuf::int32 size; (void)size;
+  int depth; (void)depth;
+  ::google::protobuf::uint32 tag;
+  ::google::protobuf::internal::ParseFunc parser_till_end; (void)parser_till_end;
+  auto ptr = begin;
+  while (ptr < end) {
+    ptr = ::google::protobuf::io::Parse32(ptr, &tag);
+    GOOGLE_PROTOBUF_PARSER_ASSERT(ptr);
+    switch (tag >> 3) {
+      // repeated int64 ids = 1;
+      case 1: {
+        if (static_cast<::google::protobuf::uint8>(tag) == 10) {
+          ptr = ::google::protobuf::io::ReadSize(ptr, &size);
+          GOOGLE_PROTOBUF_PARSER_ASSERT(ptr);
+          parser_till_end = ::google::protobuf::internal::PackedInt64Parser;
+          object = msg->mutable_ids();
+          if (size > end - ptr) goto len_delim_till_end;
+          auto newend = ptr + size;
+          if (size) ptr = parser_till_end(ptr, newend, object, ctx);
+          GOOGLE_PROTOBUF_PARSER_ASSERT(ptr == newend);
+          break;
+        } else if (static_cast<::google::protobuf::uint8>(tag) != 8) goto handle_unusual;
+        do {
+          msg->add_ids(::google::protobuf::internal::ReadVarint(&ptr));
+          GOOGLE_PROTOBUF_PARSER_ASSERT(ptr);
+          if (ptr >= end) break;
+        } while ((::google::protobuf::io::UnalignedLoad<::google::protobuf::uint64>(ptr) & 255) == 8 && (ptr += 1));
+        break;
+      }
+      // repeated string strings = 2;
+      case 2: {
+        if (static_cast<::google::protobuf::uint8>(tag) != 18) goto handle_unusual;
+        do {
+          ptr = ::google::protobuf::io::ReadSize(ptr, &size);
+          GOOGLE_PROTOBUF_PARSER_ASSERT(ptr);
+          ctx->extra_parse_data().SetFieldName("protobuf_test.Record.strings");
+          object = msg->add_strings();
+          if (size > end - ptr + ::google::protobuf::internal::ParseContext::kSlopBytes) {
+            parser_till_end = ::google::protobuf::internal::GreedyStringParserUTF8;
+            goto string_till_end;
+          }
+          GOOGLE_PROTOBUF_PARSER_ASSERT(::google::protobuf::internal::StringCheckUTF8(ptr, size, ctx));
+          ::google::protobuf::internal::InlineGreedyStringParser(object, ptr, size, ctx);
+          ptr += size;
+          if (ptr >= end) break;
+        } while ((::google::protobuf::io::UnalignedLoad<::google::protobuf::uint64>(ptr) & 255) == 18 && (ptr += 1));
+        break;
+      }
+      default: {
+      handle_unusual:
+        if ((tag & 7) == 4 || tag == 0) {
+          ctx->EndGroup(tag);
+          return ptr;
+        }
+        auto res = UnknownFieldParse(tag, {_InternalParse, msg},
+          ptr, end, msg->_internal_metadata_.mutable_unknown_fields(), ctx);
+        ptr = res.first;
+        GOOGLE_PROTOBUF_PARSER_ASSERT(ptr != nullptr);
+        if (res.second) return ptr;
+      }
+    }  // switch
+  }  // while
+  return ptr;
+string_till_end:
+  static_cast<::std::string*>(object)->clear();
+  static_cast<::std::string*>(object)->reserve(size);
+  goto len_delim_till_end;
+len_delim_till_end:
+  return ctx->StoreAndTailCall(ptr, end, {_InternalParse, msg},
+                               {parser_till_end, object}, size);
+}
+#else  // GOOGLE_PROTOBUF_ENABLE_EXPERIMENTAL_PARSER
 bool Record::MergePartialFromCodedStream(
     ::google::protobuf::io::CodedInputStream* input) {
-#define DO_(EXPRESSION) if (!GOOGLE_PREDICT_TRUE(EXPRESSION)) goto failure
+#define DO_(EXPRESSION) if (!PROTOBUF_PREDICT_TRUE(EXPRESSION)) goto failure
   ::google::protobuf::uint32 tag;
   // @@protoc_insertion_point(parse_start:protobuf_test.Record)
   for (;;) {
-    ::std::pair< ::google::protobuf::uint32, bool> p = input->ReadTagWithCutoff(127);
+    ::std::pair<::google::protobuf::uint32, bool> p = input->ReadTagWithCutoffNoLastTag(127u);
     tag = p.first;
     if (!p.second) goto handle_unusual;
     switch (::google::protobuf::internal::WireFormatLite::GetTagFieldNumber(tag)) {
       // repeated int64 ids = 1;
       case 1: {
-        if (tag == 10) {
+        if (static_cast< ::google::protobuf::uint8>(tag) == (10 & 0xFF)) {
           DO_((::google::protobuf::internal::WireFormatLite::ReadPackedPrimitive<
                    ::google::protobuf::int64, ::google::protobuf::internal::WireFormatLite::TYPE_INT64>(
                  input, this->mutable_ids())));
-        } else if (tag == 8) {
+        } else if (static_cast< ::google::protobuf::uint8>(tag) == (8 & 0xFF)) {
           DO_((::google::protobuf::internal::WireFormatLite::ReadRepeatedPrimitiveNoInline<
                    ::google::protobuf::int64, ::google::protobuf::internal::WireFormatLite::TYPE_INT64>(
-                 1, 10, input, this->mutable_ids())));
+                 1, 10u, input, this->mutable_ids())));
         } else {
           goto handle_unusual;
         }
-        if (input->ExpectTag(18)) goto parse_strings;
         break;
       }
 
       // repeated string strings = 2;
       case 2: {
-        if (tag == 18) {
-         parse_strings:
+        if (static_cast< ::google::protobuf::uint8>(tag) == (18 & 0xFF)) {
           DO_(::google::protobuf::internal::WireFormatLite::ReadString(
                 input, this->add_strings()));
           DO_(::google::protobuf::internal::WireFormatLite::VerifyUtf8String(
             this->strings(this->strings_size() - 1).data(),
-            this->strings(this->strings_size() - 1).length(),
+            static_cast<int>(this->strings(this->strings_size() - 1).length()),
             ::google::protobuf::internal::WireFormatLite::PARSE,
             "protobuf_test.Record.strings"));
         } else {
           goto handle_unusual;
         }
-        if (input->ExpectTag(18)) goto parse_strings;
-        if (input->ExpectAtEnd()) goto success;
         break;
       }
 
       default: {
       handle_unusual:
-        if (tag == 0 ||
-            ::google::protobuf::internal::WireFormatLite::GetTagWireType(tag) ==
-            ::google::protobuf::internal::WireFormatLite::WIRETYPE_END_GROUP) {
+        if (tag == 0) {
           goto success;
         }
-        DO_(::google::protobuf::internal::WireFormatLite::SkipField(input, tag));
+        DO_(::google::protobuf::internal::WireFormat::SkipField(
+              input, tag, _internal_metadata_.mutable_unknown_fields()));
         break;
       }
     }
@@ -255,37 +285,48 @@ failure:
   return false;
 #undef DO_
 }
+#endif  // GOOGLE_PROTOBUF_ENABLE_EXPERIMENTAL_PARSER
 
 void Record::SerializeWithCachedSizes(
     ::google::protobuf::io::CodedOutputStream* output) const {
   // @@protoc_insertion_point(serialize_start:protobuf_test.Record)
+  ::google::protobuf::uint32 cached_has_bits = 0;
+  (void) cached_has_bits;
+
   // repeated int64 ids = 1;
   if (this->ids_size() > 0) {
     ::google::protobuf::internal::WireFormatLite::WriteTag(1, ::google::protobuf::internal::WireFormatLite::WIRETYPE_LENGTH_DELIMITED, output);
-    output->WriteVarint32(_ids_cached_byte_size_);
+    output->WriteVarint32(_ids_cached_byte_size_.load(
+        std::memory_order_relaxed));
   }
-  for (int i = 0; i < this->ids_size(); i++) {
+  for (int i = 0, n = this->ids_size(); i < n; i++) {
     ::google::protobuf::internal::WireFormatLite::WriteInt64NoTag(
       this->ids(i), output);
   }
 
   // repeated string strings = 2;
-  for (int i = 0; i < this->strings_size(); i++) {
+  for (int i = 0, n = this->strings_size(); i < n; i++) {
     ::google::protobuf::internal::WireFormatLite::VerifyUtf8String(
-      this->strings(i).data(), this->strings(i).length(),
+      this->strings(i).data(), static_cast<int>(this->strings(i).length()),
       ::google::protobuf::internal::WireFormatLite::SERIALIZE,
       "protobuf_test.Record.strings");
     ::google::protobuf::internal::WireFormatLite::WriteString(
       2, this->strings(i), output);
   }
 
+  if (_internal_metadata_.have_unknown_fields()) {
+    ::google::protobuf::internal::WireFormat::SerializeUnknownFields(
+        _internal_metadata_.unknown_fields(), output);
+  }
   // @@protoc_insertion_point(serialize_end:protobuf_test.Record)
 }
 
 ::google::protobuf::uint8* Record::InternalSerializeWithCachedSizesToArray(
-    bool deterministic, ::google::protobuf::uint8* target) const {
-  (void)deterministic; // Unused
+    ::google::protobuf::uint8* target) const {
   // @@protoc_insertion_point(serialize_to_array_start:protobuf_test.Record)
+  ::google::protobuf::uint32 cached_has_bits = 0;
+  (void) cached_has_bits;
+
   // repeated int64 ids = 1;
   if (this->ids_size() > 0) {
     target = ::google::protobuf::internal::WireFormatLite::WriteTagToArray(
@@ -293,23 +334,26 @@ void Record::SerializeWithCachedSizes(
       ::google::protobuf::internal::WireFormatLite::WIRETYPE_LENGTH_DELIMITED,
       target);
     target = ::google::protobuf::io::CodedOutputStream::WriteVarint32ToArray(
-      _ids_cached_byte_size_, target);
-  }
-  for (int i = 0; i < this->ids_size(); i++) {
+        _ids_cached_byte_size_.load(std::memory_order_relaxed),
+         target);
     target = ::google::protobuf::internal::WireFormatLite::
-      WriteInt64NoTagToArray(this->ids(i), target);
+      WriteInt64NoTagToArray(this->ids_, target);
   }
 
   // repeated string strings = 2;
-  for (int i = 0; i < this->strings_size(); i++) {
+  for (int i = 0, n = this->strings_size(); i < n; i++) {
     ::google::protobuf::internal::WireFormatLite::VerifyUtf8String(
-      this->strings(i).data(), this->strings(i).length(),
+      this->strings(i).data(), static_cast<int>(this->strings(i).length()),
       ::google::protobuf::internal::WireFormatLite::SERIALIZE,
       "protobuf_test.Record.strings");
     target = ::google::protobuf::internal::WireFormatLite::
       WriteStringToArray(2, this->strings(i), target);
   }
 
+  if (_internal_metadata_.have_unknown_fields()) {
+    target = ::google::protobuf::internal::WireFormat::SerializeUnknownFieldsToArray(
+        _internal_metadata_.unknown_fields(), target);
+  }
   // @@protoc_insertion_point(serialize_to_array_end:protobuf_test.Record)
   return target;
 }
@@ -318,68 +362,67 @@ size_t Record::ByteSizeLong() const {
 // @@protoc_insertion_point(message_byte_size_start:protobuf_test.Record)
   size_t total_size = 0;
 
+  if (_internal_metadata_.have_unknown_fields()) {
+    total_size +=
+      ::google::protobuf::internal::WireFormat::ComputeUnknownFieldsSize(
+        _internal_metadata_.unknown_fields());
+  }
+  ::google::protobuf::uint32 cached_has_bits = 0;
+  // Prevent compiler warnings about cached_has_bits being unused
+  (void) cached_has_bits;
+
   // repeated int64 ids = 1;
   {
-    size_t data_size = 0;
-    unsigned int count = this->ids_size();
-    for (unsigned int i = 0; i < count; i++) {
-      data_size += ::google::protobuf::internal::WireFormatLite::
-        Int64Size(this->ids(i));
-    }
+    size_t data_size = ::google::protobuf::internal::WireFormatLite::
+      Int64Size(this->ids_);
     if (data_size > 0) {
       total_size += 1 +
-        ::google::protobuf::internal::WireFormatLite::Int32Size(data_size);
+        ::google::protobuf::internal::WireFormatLite::Int32Size(
+            static_cast<::google::protobuf::int32>(data_size));
     }
     int cached_size = ::google::protobuf::internal::ToCachedSize(data_size);
-    GOOGLE_SAFE_CONCURRENT_WRITES_BEGIN();
-    _ids_cached_byte_size_ = cached_size;
-    GOOGLE_SAFE_CONCURRENT_WRITES_END();
+    _ids_cached_byte_size_.store(cached_size,
+                                    std::memory_order_relaxed);
     total_size += data_size;
   }
 
   // repeated string strings = 2;
   total_size += 1 *
       ::google::protobuf::internal::FromIntSize(this->strings_size());
-  for (int i = 0; i < this->strings_size(); i++) {
+  for (int i = 0, n = this->strings_size(); i < n; i++) {
     total_size += ::google::protobuf::internal::WireFormatLite::StringSize(
       this->strings(i));
   }
 
   int cached_size = ::google::protobuf::internal::ToCachedSize(total_size);
-  GOOGLE_SAFE_CONCURRENT_WRITES_BEGIN();
-  _cached_size_ = cached_size;
-  GOOGLE_SAFE_CONCURRENT_WRITES_END();
+  SetCachedSize(cached_size);
   return total_size;
 }
 
 void Record::MergeFrom(const ::google::protobuf::Message& from) {
 // @@protoc_insertion_point(generalized_merge_from_start:protobuf_test.Record)
-  if (GOOGLE_PREDICT_FALSE(&from == this)) MergeFromFail(__LINE__);
+  GOOGLE_DCHECK_NE(&from, this);
   const Record* source =
-      ::google::protobuf::internal::DynamicCastToGenerated<const Record>(
+      ::google::protobuf::DynamicCastToGenerated<Record>(
           &from);
-  if (source == NULL) {
+  if (source == nullptr) {
   // @@protoc_insertion_point(generalized_merge_from_cast_fail:protobuf_test.Record)
     ::google::protobuf::internal::ReflectionOps::Merge(from, this);
   } else {
   // @@protoc_insertion_point(generalized_merge_from_cast_success:protobuf_test.Record)
-    UnsafeMergeFrom(*source);
+    MergeFrom(*source);
   }
 }
 
 void Record::MergeFrom(const Record& from) {
 // @@protoc_insertion_point(class_specific_merge_from_start:protobuf_test.Record)
-  if (GOOGLE_PREDICT_TRUE(&from != this)) {
-    UnsafeMergeFrom(from);
-  } else {
-    MergeFromFail(__LINE__);
-  }
-}
+  GOOGLE_DCHECK_NE(&from, this);
+  _internal_metadata_.MergeFrom(from._internal_metadata_);
+  ::google::protobuf::uint32 cached_has_bits = 0;
+  (void) cached_has_bits;
 
-void Record::UnsafeMergeFrom(const Record& from) {
-  GOOGLE_DCHECK(&from != this);
-  ids_.UnsafeMergeFrom(from.ids_);
-  strings_.UnsafeMergeFrom(from.strings_);
+  ids_.MergeFrom(from.ids_);
+  strings_.MergeFrom(from.strings_);
 }
 
 void Record::CopyFrom(const ::google::protobuf::Message& from) {
@@ -393,11 +436,10 @@ void Record::CopyFrom(const Record& from) {
 // @@protoc_insertion_point(class_specific_copy_from_start:protobuf_test.Record)
   if (&from == this) return;
   Clear();
-  UnsafeMergeFrom(from);
+  MergeFrom(from);
 }
 
 bool Record::IsInitialized() const {
-
   return true;
 }
 
@@ -406,115 +448,27 @@ void Record::Swap(Record* other) {
   InternalSwap(other);
 }
 void Record::InternalSwap(Record* other) {
-  ids_.UnsafeArenaSwap(&other->ids_);
-  strings_.UnsafeArenaSwap(&other->strings_);
+  using std::swap;
   _internal_metadata_.Swap(&other->_internal_metadata_);
-  std::swap(_cached_size_, other->_cached_size_);
+  ids_.InternalSwap(&other->ids_);
+  strings_.InternalSwap(CastToBase(&other->strings_));
 }
 
 ::google::protobuf::Metadata Record::GetMetadata() const {
-  protobuf_AssignDescriptorsOnce();
-  ::google::protobuf::Metadata metadata;
-  metadata.descriptor = Record_descriptor_;
-  metadata.reflection = Record_reflection_;
-  return metadata;
+  ::google::protobuf::internal::AssignDescriptors(&::assign_descriptors_table_test_2eproto);
+  return ::file_level_metadata_test_2eproto[kIndexInFileMessages];
 }
 
-#if PROTOBUF_INLINE_NOT_IN_HEADERS
-// Record
-
-// repeated int64 ids = 1;
-int Record::ids_size() const {
-  return ids_.size();
-}
-void Record::clear_ids() {
-  ids_.Clear();
-}
-::google::protobuf::int64 Record::ids(int index) const {
-  // @@protoc_insertion_point(field_get:protobuf_test.Record.ids)
-  return ids_.Get(index);
-}
-void Record::set_ids(int index, ::google::protobuf::int64 value) {
-  ids_.Set(index, value);
-  // @@protoc_insertion_point(field_set:protobuf_test.Record.ids)
-}
-void Record::add_ids(::google::protobuf::int64 value) {
-  ids_.Add(value);
-  // @@protoc_insertion_point(field_add:protobuf_test.Record.ids)
-}
-const ::google::protobuf::RepeatedField< ::google::protobuf::int64 >&
-Record::ids() const {
-  // @@protoc_insertion_point(field_list:protobuf_test.Record.ids)
-  return ids_;
-}
-::google::protobuf::RepeatedField< ::google::protobuf::int64 >*
-Record::mutable_ids() {
-  // @@protoc_insertion_point(field_mutable_list:protobuf_test.Record.ids)
-  return &ids_;
-}
-
-// repeated string strings = 2;
-int Record::strings_size() const {
-  return strings_.size();
-}
-void Record::clear_strings() {
-  strings_.Clear();
-}
-const ::std::string& Record::strings(int index) const {
-  // @@protoc_insertion_point(field_get:protobuf_test.Record.strings)
-  return strings_.Get(index);
-}
-::std::string* Record::mutable_strings(int index) {
-  // @@protoc_insertion_point(field_mutable:protobuf_test.Record.strings)
-  return strings_.Mutable(index);
-}
-void Record::set_strings(int index, const ::std::string& value) {
-  // @@protoc_insertion_point(field_set:protobuf_test.Record.strings)
-  strings_.Mutable(index)->assign(value);
-}
-void Record::set_strings(int index, const char* value) {
-  strings_.Mutable(index)->assign(value);
-  // @@protoc_insertion_point(field_set_char:protobuf_test.Record.strings)
-}
-void Record::set_strings(int index, const char* value, size_t size) {
-  strings_.Mutable(index)->assign(
-    reinterpret_cast<const char*>(value), size);
-  // @@protoc_insertion_point(field_set_pointer:protobuf_test.Record.strings)
-}
-::std::string* Record::add_strings() {
-  // @@protoc_insertion_point(field_add_mutable:protobuf_test.Record.strings)
-  return strings_.Add();
-}
-void Record::add_strings(const ::std::string& value) {
-  strings_.Add()->assign(value);
-  // @@protoc_insertion_point(field_add:protobuf_test.Record.strings)
-}
-void Record::add_strings(const char* value) {
-  strings_.Add()->assign(value);
-  // @@protoc_insertion_point(field_add_char:protobuf_test.Record.strings)
-}
-void Record::add_strings(const char* value, size_t size) {
-  strings_.Add()->assign(reinterpret_cast<const char*>(value), size);
-  // @@protoc_insertion_point(field_add_pointer:protobuf_test.Record.strings)
-}
-const ::google::protobuf::RepeatedPtrField< ::std::string>&
-Record::strings() const {
-  // @@protoc_insertion_point(field_list:protobuf_test.Record.strings)
-  return strings_;
-}
-::google::protobuf::RepeatedPtrField< ::std::string>*
-Record::mutable_strings() {
-  // @@protoc_insertion_point(field_mutable_list:protobuf_test.Record.strings)
-  return &strings_;
-}
-
-inline const Record* Record::internal_default_instance() {
-  return &Record_default_instance_.get();
-}
-#endif  // PROTOBUF_INLINE_NOT_IN_HEADERS
 
 // @@protoc_insertion_point(namespace_scope)
-
 }  // namespace protobuf_test
+namespace google {
+namespace protobuf {
+template<> PROTOBUF_NOINLINE ::protobuf_test::Record* Arena::CreateMaybeMessage< ::protobuf_test::Record >(Arena* arena) {
+  return Arena::CreateInternal< ::protobuf_test::Record >(arena);
+}
+}  // namespace protobuf
+}  // namespace google
 
 // @@protoc_insertion_point(global_scope)
+#include <google/protobuf/port_undef.inc>
