@@ -1,47 +1,69 @@
 // Generated by the protocol buffer compiler.  DO NOT EDIT!
 // source: test.proto
 
-#ifndef PROTOBUF_test_2eproto__INCLUDED
-#define PROTOBUF_test_2eproto__INCLUDED
+#ifndef PROTOBUF_INCLUDED_test_2eproto
+#define PROTOBUF_INCLUDED_test_2eproto
 
+#include <limits>
 #include <string>
 
-#include <google/protobuf/stubs/common.h>
-
-#if GOOGLE_PROTOBUF_VERSION < 3001000
+#include <google/protobuf/port_def.inc>
+#if PROTOBUF_VERSION < 3007000
 #error This file was generated by a newer version of protoc which is
-#error incompatible with your Protocol Buffer headers.  Please update
+#error incompatible with your Protocol Buffer headers. Please update
 #error your headers.
 #endif
-#if 3001000 < GOOGLE_PROTOBUF_MIN_PROTOC_VERSION
+#if 3007000 < PROTOBUF_MIN_PROTOC_VERSION
 #error This file was generated by an older version of protoc which is
-#error incompatible with your Protocol Buffer headers.  Please
+#error incompatible with your Protocol Buffer headers. Please
 #error regenerate this file with a newer version of protoc.
 #endif
 
+#include <google/protobuf/port_undef.inc>
+#include <google/protobuf/io/coded_stream.h>
 #include <google/protobuf/arena.h>
 #include <google/protobuf/arenastring.h>
+#include <google/protobuf/generated_message_table_driven.h>
 #include <google/protobuf/generated_message_util.h>
+#include <google/protobuf/inlined_string_field.h>
 #include <google/protobuf/metadata.h>
 #include <google/protobuf/message.h>
-#include <google/protobuf/repeated_field.h>
-#include <google/protobuf/extension_set.h>
+#include <google/protobuf/repeated_field.h>  // IWYU pragma: export
+#include <google/protobuf/extension_set.h>  // IWYU pragma: export
 #include <google/protobuf/unknown_field_set.h>
 // @@protoc_insertion_point(includes)
+#include <google/protobuf/port_def.inc>
+#define PROTOBUF_INTERNAL_EXPORT_test_2eproto
 
+// Internal implementation detail -- do not use these members.
+struct TableStruct_test_2eproto {
+  static const ::google::protobuf::internal::ParseTableField entries[]
+    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
+  static const ::google::protobuf::internal::AuxillaryParseTableField aux[]
+    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
+  static const ::google::protobuf::internal::ParseTable schema[1]
+    PROTOBUF_SECTION_VARIABLE(protodesc_cold);
+  static const ::google::protobuf::internal::FieldMetadata field_metadata[];
+  static const ::google::protobuf::internal::SerializationTable serialization_table[];
+  static const ::google::protobuf::uint32 offsets[];
+};
+void AddDescriptors_test_2eproto();
 namespace protobuf_test {
-
-// Internal implementation detail -- do not call these.
-void protobuf_AddDesc_test_2eproto();
-void protobuf_InitDefaults_test_2eproto();
-void protobuf_AssignDesc_test_2eproto();
-void protobuf_ShutdownFile_test_2eproto();
-
 class Record;
+class RecordDefaultTypeInternal;
+extern RecordDefaultTypeInternal _Record_default_instance_;
+}  // namespace protobuf_test
+namespace google {
+namespace protobuf {
+template<> ::protobuf_test::Record* Arena::CreateMaybeMessage<::protobuf_test::Record>(Arena*);
+}  // namespace protobuf
+}  // namespace google
+namespace protobuf_test {
 
 // ===================================================================
 
-class Record : public ::google::protobuf::Message /* @@protoc_insertion_point(class_definition:protobuf_test.Record) */ {
+class Record final :
+    public ::google::protobuf::Message /* @@protoc_insertion_point(class_definition:protobuf_test.Record) */ {
  public:
   Record();
   virtual ~Record();
@@ -52,53 +74,84 @@ class Record : public ::google::protobuf::Message /* @@protoc_insertion_point(cl
     CopyFrom(from);
     return *this;
   }
+  #if LANG_CXX11
+  Record(Record&& from) noexcept
+    : Record() {
+    *this = ::std::move(from);
+  }
 
-  static const ::google::protobuf::Descriptor* descriptor();
+  inline Record& operator=(Record&& from) noexcept {
+    if (GetArenaNoVirtual() == from.GetArenaNoVirtual()) {
+      if (this != &from) InternalSwap(&from);
+    } else {
+      CopyFrom(from);
+    }
+    return *this;
+  }
+  #endif
+  static const ::google::protobuf::Descriptor* descriptor() {
+    return default_instance().GetDescriptor();
+  }
   static const Record& default_instance();
 
-  static const Record* internal_default_instance();
+  static void InitAsDefaultInstance();  // FOR INTERNAL USE ONLY
+  static inline const Record* internal_default_instance() {
+    return reinterpret_cast<const Record*>(
+               &_Record_default_instance_);
+  }
+  static constexpr int kIndexInFileMessages =
+    0;
 
   void Swap(Record* other);
+  friend void swap(Record& a, Record& b) {
+    a.Swap(&b);
+  }
 
   // implements Message ----------------------------------------------
 
-  inline Record* New() const { return New(NULL); }
+  inline Record* New() const final {
+    return CreateMaybeMessage<Record>(nullptr);
+  }
 
-  Record* New(::google::protobuf::Arena* arena) const;
-  void CopyFrom(const ::google::protobuf::Message& from);
-  void MergeFrom(const ::google::protobuf::Message& from);
+  Record* New(::google::protobuf::Arena* arena) const final {
+    return CreateMaybeMessage<Record>(arena);
+  }
+  void CopyFrom(const ::google::protobuf::Message& from) final;
+  void MergeFrom(const ::google::protobuf::Message& from) final;
   void CopyFrom(const Record& from);
   void MergeFrom(const Record& from);
-  void Clear();
-  bool IsInitialized() const;
+  PROTOBUF_ATTRIBUTE_REINITIALIZES void Clear() final;
+  bool IsInitialized() const final;
 
-  size_t ByteSizeLong() const;
+  size_t ByteSizeLong() const final;
+  #if GOOGLE_PROTOBUF_ENABLE_EXPERIMENTAL_PARSER
+  static const char* _InternalParse(const char* begin, const char* end, void* object, ::google::protobuf::internal::ParseContext* ctx);
+  ::google::protobuf::internal::ParseFunc _ParseFunc() const final { return _InternalParse; }
+  #else
   bool MergePartialFromCodedStream(
-      ::google::protobuf::io::CodedInputStream* input);
+      ::google::protobuf::io::CodedInputStream* input) final;
+  #endif  // GOOGLE_PROTOBUF_ENABLE_EXPERIMENTAL_PARSER
   void SerializeWithCachedSizes(
-      ::google::protobuf::io::CodedOutputStream* output) const;
+      ::google::protobuf::io::CodedOutputStream* output) const final;
   ::google::protobuf::uint8* InternalSerializeWithCachedSizesToArray(
-      bool deterministic, ::google::protobuf::uint8* output) const;
-  ::google::protobuf::uint8* SerializeWithCachedSizesToArray(::google::protobuf::uint8* output) const {
-    return InternalSerializeWithCachedSizesToArray(false, output);
-  }
-  int GetCachedSize() const { return _cached_size_; }
+      ::google::protobuf::uint8* target) const final;
+  int GetCachedSize() const final { return _cached_size_.Get(); }
+
   private:
   void SharedCtor();
   void SharedDtor();
-  void SetCachedSize(int size) const;
+  void SetCachedSize(int size) const final;
   void InternalSwap(Record* other);
-  void UnsafeMergeFrom(const Record& from);
   private:
   inline ::google::protobuf::Arena* GetArenaNoVirtual() const {
-    return _internal_metadata_.arena();
+    return nullptr;
   }
   inline void* MaybeArenaPtr() const {
-    return _internal_metadata_.raw_arena_ptr();
+    return nullptr;
   }
   public:
 
-  ::google::protobuf::Metadata GetMetadata() const;
+  ::google::protobuf::Metadata GetMetadata() const final;
 
   // nested types ----------------------------------------------------
 
@@ -123,38 +176,41 @@ class Record : public ::google::protobuf::Message /* @@protoc_insertion_point(cl
   const ::std::string& strings(int index) const;
   ::std::string* mutable_strings(int index);
   void set_strings(int index, const ::std::string& value);
+  #if LANG_CXX11
+  void set_strings(int index, ::std::string&& value);
+  #endif
   void set_strings(int index, const char* value);
   void set_strings(int index, const char* value, size_t size);
   ::std::string* add_strings();
   void add_strings(const ::std::string& value);
+  #if LANG_CXX11
+  void add_strings(::std::string&& value);
+  #endif
   void add_strings(const char* value);
   void add_strings(const char* value, size_t size);
-  const ::google::protobuf::RepeatedPtrField< ::std::string>& strings() const;
-  ::google::protobuf::RepeatedPtrField< ::std::string>* mutable_strings();
+  const ::google::protobuf::RepeatedPtrField<::std::string>& strings() const;
+  ::google::protobuf::RepeatedPtrField<::std::string>* mutable_strings();
 
   // @@protoc_insertion_point(class_scope:protobuf_test.Record)
  private:
+  class HasBitSetters;
 
   ::google::protobuf::internal::InternalMetadataWithArena _internal_metadata_;
   ::google::protobuf::RepeatedField< ::google::protobuf::int64 > ids_;
-  mutable int _ids_cached_byte_size_;
-  ::google::protobuf::RepeatedPtrField< ::std::string> strings_;
-  mutable int _cached_size_;
-  friend void  protobuf_InitDefaults_test_2eproto_impl();
-  friend void  protobuf_AddDesc_test_2eproto_impl();
-  friend void protobuf_AssignDesc_test_2eproto();
-  friend void protobuf_ShutdownFile_test_2eproto();
-
-  void InitAsDefaultInstance();
+  mutable std::atomic<int> _ids_cached_byte_size_;
+  ::google::protobuf::RepeatedPtrField<::std::string> strings_;
+  mutable ::google::protobuf::internal::CachedSize _cached_size_;
+  friend struct ::TableStruct_test_2eproto;
 };
-extern ::google::protobuf::internal::ExplicitlyConstructed<Record> Record_default_instance_;
-
 // ===================================================================
 
 
 // ===================================================================
 
-#if !PROTOBUF_INLINE_NOT_IN_HEADERS
+#ifdef __GNUC__
+  #pragma GCC diagnostic push
+  #pragma GCC diagnostic ignored "-Wstrict-aliasing"
+#endif  // __GNUC__
 // Record
 
 // repeated int64 ids = 1;
@@ -206,7 +262,14 @@ inline void Record::set_strings(int index, const ::std::string& value) {
   // @@protoc_insertion_point(field_set:protobuf_test.Record.strings)
   strings_.Mutable(index)->assign(value);
 }
+#if LANG_CXX11
+inline void Record::set_strings(int index, ::std::string&& value) {
+  // @@protoc_insertion_point(field_set:protobuf_test.Record.strings)
+  strings_.Mutable(index)->assign(std::move(value));
+}
+#endif
 inline void Record::set_strings(int index, const char* value) {
+  GOOGLE_DCHECK(value != nullptr);
   strings_.Mutable(index)->assign(value);
   // @@protoc_insertion_point(field_set_char:protobuf_test.Record.strings)
 }
@@ -223,7 +286,14 @@ inline void Record::add_strings(const ::std::string& value) {
   strings_.Add()->assign(value);
   // @@protoc_insertion_point(field_add:protobuf_test.Record.strings)
 }
+#if LANG_CXX11
+inline void Record::add_strings(::std::string&& value) {
+  strings_.Add(std::move(value));
+  // @@protoc_insertion_point(field_add:protobuf_test.Record.strings)
+}
+#endif
 inline void Record::add_strings(const char* value) {
+  GOOGLE_DCHECK(value != nullptr);
   strings_.Add()->assign(value);
   // @@protoc_insertion_point(field_add_char:protobuf_test.Record.strings)
 }
@@ -231,21 +301,20 @@ inline void Record::add_strings(const char* value, size_t size) {
   strings_.Add()->assign(reinterpret_cast<const char*>(value), size);
   // @@protoc_insertion_point(field_add_pointer:protobuf_test.Record.strings)
 }
-inline const ::google::protobuf::RepeatedPtrField< ::std::string>&
+inline const ::google::protobuf::RepeatedPtrField<::std::string>&
 Record::strings() const {
   // @@protoc_insertion_point(field_list:protobuf_test.Record.strings)
   return strings_;
 }
-inline ::google::protobuf::RepeatedPtrField< ::std::string>*
+inline ::google::protobuf::RepeatedPtrField<::std::string>*
 Record::mutable_strings() {
   // @@protoc_insertion_point(field_mutable_list:protobuf_test.Record.strings)
   return &strings_;
 }
 
-inline const Record* Record::internal_default_instance() {
-  return &Record_default_instance_.get();
-}
-#endif  // !PROTOBUF_INLINE_NOT_IN_HEADERS
+#ifdef __GNUC__
+  #pragma GCC diagnostic pop
+#endif  // __GNUC__
 
 // @@protoc_insertion_point(namespace_scope)
 
@@ -253,4 +322,5 @@ inline const Record* Record::internal_default_instance() {
 
 // @@protoc_insertion_point(global_scope)
 
-#endif  // PROTOBUF_test_2eproto__INCLUDED
+#include <google/protobuf/port_undef.inc>
+#endif  // PROTOBUF_INCLUDED_test_2eproto
