@@ -17,8 +17,8 @@
  */
 
 
-#ifndef _HOME_NIXMAN_4_CPP_SERIALIZERS_AVRO_RECORD_HPP_3019802276__H_
-#define _HOME_NIXMAN_4_CPP_SERIALIZERS_AVRO_RECORD_HPP_3019802276__H_
+#ifndef _HOME_PKOK_CPP_SERIALIZERS_AVRO_RECORD_HPP_2710640864__H_
+#define _HOME_PKOK_CPP_SERIALIZERS_AVRO_RECORD_HPP_2710640864__H_
 
 
 #include <sstream>
