@@ -1,12 +1,12 @@
 // Generated by Cap'n Proto compiler, DO NOT EDIT
 // source: test.capnp
 
-#ifndef CAPNP_INCLUDED_ead5e4c3b2579756_
-#define CAPNP_INCLUDED_ead5e4c3b2579756_
+#pragma once
 
 #include <capnp/generated-header-support.h>
+#include <kj/windows-sanity.h>
 
-#if CAPNP_VERSION != 6001
+#if CAPNP_VERSION != 7000
 #error "Version mismatch between generated code and library headers.  You must use the same version of the Cap'n Proto compiler and library."
 #endif
 
@@ -56,10 +56,10 @@ public:
 #endif  // !CAPNP_LITE
 
   inline bool hasIds() const;
-  inline  ::capnp::List< ::int64_t>::Reader getIds() const;
+  inline  ::capnp::List< ::int64_t,  ::capnp::Kind::PRIMITIVE>::Reader getIds() const;
 
   inline bool hasStrings() const;
-  inline  ::capnp::List< ::capnp::Text>::Reader getStrings() const;
+  inline  ::capnp::List< ::capnp::Text,  ::capnp::Kind::BLOB>::Reader getStrings() const;
 
 private:
   ::capnp::_::StructReader _reader;
@@ -90,20 +90,20 @@ public:
 #endif  // !CAPNP_LITE
 
   inline bool hasIds();
-  inline  ::capnp::List< ::int64_t>::Builder getIds();
-  inline void setIds( ::capnp::List< ::int64_t>::Reader value);
+  inline  ::capnp::List< ::int64_t,  ::capnp::Kind::PRIMITIVE>::Builder getIds();
+  inline void setIds( ::capnp::List< ::int64_t,  ::capnp::Kind::PRIMITIVE>::Reader value);
   inline void setIds(::kj::ArrayPtr<const  ::int64_t> value);
-  inline  ::capnp::List< ::int64_t>::Builder initIds(unsigned int size);
-  inline void adoptIds(::capnp::Orphan< ::capnp::List< ::int64_t>>&& value);
-  inline ::capnp::Orphan< ::capnp::List< ::int64_t>> disownIds();
+  inline  ::capnp::List< ::int64_t,  ::capnp::Kind::PRIMITIVE>::Builder initIds(unsigned int size);
+  inline void adoptIds(::capnp::Orphan< ::capnp::List< ::int64_t,  ::capnp::Kind::PRIMITIVE>>&& value);
+  inline ::capnp::Orphan< ::capnp::List< ::int64_t,  ::capnp::Kind::PRIMITIVE>> disownIds();
 
   inline bool hasStrings();
-  inline  ::capnp::List< ::capnp::Text>::Builder getStrings();
-  inline void setStrings( ::capnp::List< ::capnp::Text>::Reader value);
+  inline  ::capnp::List< ::capnp::Text,  ::capnp::Kind::BLOB>::Builder getStrings();
+  inline void setStrings( ::capnp::List< ::capnp::Text,  ::capnp::Kind::BLOB>::Reader value);
   inline void setStrings(::kj::ArrayPtr<const  ::capnp::Text::Reader> value);
-  inline  ::capnp::List< ::capnp::Text>::Builder initStrings(unsigned int size);
-  inline void adoptStrings(::capnp::Orphan< ::capnp::List< ::capnp::Text>>&& value);
-  inline ::capnp::Orphan< ::capnp::List< ::capnp::Text>> disownStrings();
+  inline  ::capnp::List< ::capnp::Text,  ::capnp::Kind::BLOB>::Builder initStrings(unsigned int size);
+  inline void adoptStrings(::capnp::Orphan< ::capnp::List< ::capnp::Text,  ::capnp::Kind::BLOB>>&& value);
+  inline ::capnp::Orphan< ::capnp::List< ::capnp::Text,  ::capnp::Kind::BLOB>> disownStrings();
 
 private:
   ::capnp::_::StructBuilder _builder;
@@ -141,33 +141,33 @@ inline bool Record::Builder::hasIds() {
   return !_builder.getPointerField(
       ::capnp::bounded<0>() * ::capnp::POINTERS).isNull();
 }
-inline  ::capnp::List< ::int64_t>::Reader Record::Reader::getIds() const {
-  return ::capnp::_::PointerHelpers< ::capnp::List< ::int64_t>>::get(_reader.getPointerField(
+inline  ::capnp::List< ::int64_t,  ::capnp::Kind::PRIMITIVE>::Reader Record::Reader::getIds() const {
+  return ::capnp::_::PointerHelpers< ::capnp::List< ::int64_t,  ::capnp::Kind::PRIMITIVE>>::get(_reader.getPointerField(
       ::capnp::bounded<0>() * ::capnp::POINTERS));
 }
-inline  ::capnp::List< ::int64_t>::Builder Record::Builder::getIds() {
-  return ::capnp::_::PointerHelpers< ::capnp::List< ::int64_t>>::get(_builder.getPointerField(
+inline  ::capnp::List< ::int64_t,  ::capnp::Kind::PRIMITIVE>::Builder Record::Builder::getIds() {
+  return ::capnp::_::PointerHelpers< ::capnp::List< ::int64_t,  ::capnp::Kind::PRIMITIVE>>::get(_builder.getPointerField(
       ::capnp::bounded<0>() * ::capnp::POINTERS));
 }
-inline void Record::Builder::setIds( ::capnp::List< ::int64_t>::Reader value) {
-  ::capnp::_::PointerHelpers< ::capnp::List< ::int64_t>>::set(_builder.getPointerField(
+inline void Record::Builder::setIds( ::capnp::List< ::int64_t,  ::capnp::Kind::PRIMITIVE>::Reader value) {
+  ::capnp::_::PointerHelpers< ::capnp::List< ::int64_t,  ::capnp::Kind::PRIMITIVE>>::set(_builder.getPointerField(
       ::capnp::bounded<0>() * ::capnp::POINTERS), value);
 }
 inline void Record::Builder::setIds(::kj::ArrayPtr<const  ::int64_t> value) {
-  ::capnp::_::PointerHelpers< ::capnp::List< ::int64_t>>::set(_builder.getPointerField(
+  ::capnp::_::PointerHelpers< ::capnp::List< ::int64_t,  ::capnp::Kind::PRIMITIVE>>::set(_builder.getPointerField(
       ::capnp::bounded<0>() * ::capnp::POINTERS), value);
 }
-inline  ::capnp::List< ::int64_t>::Builder Record::Builder::initIds(unsigned int size) {
-  return ::capnp::_::PointerHelpers< ::capnp::List< ::int64_t>>::init(_builder.getPointerField(
+inline  ::capnp::List< ::int64_t,  ::capnp::Kind::PRIMITIVE>::Builder Record::Builder::initIds(unsigned int size) {
+  return ::capnp::_::PointerHelpers< ::capnp::List< ::int64_t,  ::capnp::Kind::PRIMITIVE>>::init(_builder.getPointerField(
       ::capnp::bounded<0>() * ::capnp::POINTERS), size);
 }
 inline void Record::Builder::adoptIds(
-    ::capnp::Orphan< ::capnp::List< ::int64_t>>&& value) {
-  ::capnp::_::PointerHelpers< ::capnp::List< ::int64_t>>::adopt(_builder.getPointerField(
+    ::capnp::Orphan< ::capnp::List< ::int64_t,  ::capnp::Kind::PRIMITIVE>>&& value) {
+  ::capnp::_::PointerHelpers< ::capnp::List< ::int64_t,  ::capnp::Kind::PRIMITIVE>>::adopt(_builder.getPointerField(
       ::capnp::bounded<0>() * ::capnp::POINTERS), kj::mv(value));
 }
-inline ::capnp::Orphan< ::capnp::List< ::int64_t>> Record::Builder::disownIds() {
-  return ::capnp::_::PointerHelpers< ::capnp::List< ::int64_t>>::disown(_builder.getPointerField(
+inline ::capnp::Orphan< ::capnp::List< ::int64_t,  ::capnp::Kind::PRIMITIVE>> Record::Builder::disownIds() {
+  return ::capnp::_::PointerHelpers< ::capnp::List< ::int64_t,  ::capnp::Kind::PRIMITIVE>>::disown(_builder.getPointerField(
       ::capnp::bounded<0>() * ::capnp::POINTERS));
 }
 
@@ -179,36 +179,35 @@ inline bool Record::Builder::hasStrings() {
   return !_builder.getPointerField(
       ::capnp::bounded<1>() * ::capnp::POINTERS).isNull();
 }
-inline  ::capnp::List< ::capnp::Text>::Reader Record::Reader::getStrings() const {
-  return ::capnp::_::PointerHelpers< ::capnp::List< ::capnp::Text>>::get(_reader.getPointerField(
+inline  ::capnp::List< ::capnp::Text,  ::capnp::Kind::BLOB>::Reader Record::Reader::getStrings() const {
+  return ::capnp::_::PointerHelpers< ::capnp::List< ::capnp::Text,  ::capnp::Kind::BLOB>>::get(_reader.getPointerField(
       ::capnp::bounded<1>() * ::capnp::POINTERS));
 }
-inline  ::capnp::List< ::capnp::Text>::Builder Record::Builder::getStrings() {
-  return ::capnp::_::PointerHelpers< ::capnp::List< ::capnp::Text>>::get(_builder.getPointerField(
+inline  ::capnp::List< ::capnp::Text,  ::capnp::Kind::BLOB>::Builder Record::Builder::getStrings() {
+  return ::capnp::_::PointerHelpers< ::capnp::List< ::capnp::Text,  ::capnp::Kind::BLOB>>::get(_builder.getPointerField(
       ::capnp::bounded<1>() * ::capnp::POINTERS));
 }
-inline void Record::Builder::setStrings( ::capnp::List< ::capnp::Text>::Reader value) {
-  ::capnp::_::PointerHelpers< ::capnp::List< ::capnp::Text>>::set(_builder.getPointerField(
+inline void Record::Builder::setStrings( ::capnp::List< ::capnp::Text,  ::capnp::Kind::BLOB>::Reader value) {
+  ::capnp::_::PointerHelpers< ::capnp::List< ::capnp::Text,  ::capnp::Kind::BLOB>>::set(_builder.getPointerField(
       ::capnp::bounded<1>() * ::capnp::POINTERS), value);
 }
 inline void Record::Builder::setStrings(::kj::ArrayPtr<const  ::capnp::Text::Reader> value) {
-  ::capnp::_::PointerHelpers< ::capnp::List< ::capnp::Text>>::set(_builder.getPointerField(
+  ::capnp::_::PointerHelpers< ::capnp::List< ::capnp::Text,  ::capnp::Kind::BLOB>>::set(_builder.getPointerField(
       ::capnp::bounded<1>() * ::capnp::POINTERS), value);
 }
-inline  ::capnp::List< ::capnp::Text>::Builder Record::Builder::initStrings(unsigned int size) {
-  return ::capnp::_::PointerHelpers< ::capnp::List< ::capnp::Text>>::init(_builder.getPointerField(
+inline  ::capnp::List< ::capnp::Text,  ::capnp::Kind::BLOB>::Builder Record::Builder::initStrings(unsigned int size) {
+  return ::capnp::_::PointerHelpers< ::capnp::List< ::capnp::Text,  ::capnp::Kind::BLOB>>::init(_builder.getPointerField(
       ::capnp::bounded<1>() * ::capnp::POINTERS), size);
 }
 inline void Record::Builder::adoptStrings(
-    ::capnp::Orphan< ::capnp::List< ::capnp::Text>>&& value) {
-  ::capnp::_::PointerHelpers< ::capnp::List< ::capnp::Text>>::adopt(_builder.getPointerField(
+    ::capnp::Orphan< ::capnp::List< ::capnp::Text,  ::capnp::Kind::BLOB>>&& value) {
+  ::capnp::_::PointerHelpers< ::capnp::List< ::capnp::Text,  ::capnp::Kind::BLOB>>::adopt(_builder.getPointerField(
       ::capnp::bounded<1>() * ::capnp::POINTERS), kj::mv(value));
 }
-inline ::capnp::Orphan< ::capnp::List< ::capnp::Text>> Record::Builder::disownStrings() {
-  return ::capnp::_::PointerHelpers< ::capnp::List< ::capnp::Text>>::disown(_builder.getPointerField(
+inline ::capnp::Orphan< ::capnp::List< ::capnp::Text,  ::capnp::Kind::BLOB>> Record::Builder::disownStrings() {
+  return ::capnp::_::PointerHelpers< ::capnp::List< ::capnp::Text,  ::capnp::Kind::BLOB>>::disown(_builder.getPointerField(
       ::capnp::bounded<1>() * ::capnp::POINTERS));
 }
 
 }  // namespace
 
-#endif  // CAPNP_INCLUDED_ead5e4c3b2579756_
