@@ -1,5 +1,5 @@
 /**
- * Autogenerated by Thrift Compiler (0.9.3)
+ * Autogenerated by Thrift Compiler (0.12.0)
  *
  * DO NOT EDIT UNLESS YOU ARE SURE THAT YOU KNOW WHAT YOU ARE DOING
  *  @generated
@@ -25,10 +25,16 @@ void Record::__set_ids(const std::vector<int64_t> & val) {
 void Record::__set_strings(const std::vector<std::string> & val) {
   this->strings = val;
 }
+std::ostream& operator<<(std::ostream& out, const Record& obj)
+{
+  obj.printTo(out);
+  return out;
+}
+
 
 uint32_t Record::read(::apache::thrift::protocol::TProtocol* iprot) {
 
-  apache::thrift::protocol::TInputRecursionTracker tracker(*iprot);
+  ::apache::thrift::protocol::TInputRecursionTracker tracker(*iprot);
   uint32_t xfer = 0;
   std::string fname;
   ::apache::thrift::protocol::TType ftype;
@@ -107,7 +113,7 @@ uint32_t Record::read(::apache::thrift::protocol::TProtocol* iprot) {
 
 uint32_t Record::write(::apache::thrift::protocol::TProtocol* oprot) const {
   uint32_t xfer = 0;
-  apache::thrift::protocol::TOutputRecursionTracker tracker(*oprot);
+  ::apache::thrift::protocol::TOutputRecursionTracker tracker(*oprot);
   xfer += oprot->writeStructBegin("Record");
 
   xfer += oprot->writeFieldBegin("ids", ::apache::thrift::protocol::T_LIST, 1);
