@@ -1,5 +1,5 @@
 /**
- * Autogenerated by Thrift Compiler (0.9.3)
+ * Autogenerated by Thrift Compiler (0.12.0)
  *
  * DO NOT EDIT UNLESS YOU ARE SURE THAT YOU KNOW WHAT YOU ARE DOING
  *  @generated
@@ -11,10 +11,11 @@
 
 #include <thrift/Thrift.h>
 #include <thrift/TApplicationException.h>
+#include <thrift/TBase.h>
 #include <thrift/protocol/TProtocol.h>
 #include <thrift/transport/TTransport.h>
 
-#include <thrift/cxxfunctional.h>
+#include <thrift/stdcxx.h>
 
 
 namespace thrift_test {
@@ -22,7 +23,7 @@ namespace thrift_test {
 class Record;
 
 
-class Record {
+class Record : public virtual ::apache::thrift::TBase {
  public:
 
   Record(const Record&);
@@ -60,11 +61,7 @@ class Record {
 
 void swap(Record &a, Record &b);
 
-inline std::ostream& operator<<(std::ostream& out, const Record& obj)
-{
-  obj.printTo(out);
-  return out;
-}
+std::ostream& operator<<(std::ostream& out, const Record& obj);
 
 } // namespace
 
